chore: move dryrunfifo
